backend/pyproject.toml
[tool.poetry]
name = "adk-backend"
version = "0.1.0"
description = "A multi-agent system with Google's ADK"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
google-agent-kit = "*"
google-cloud-aiplatform = "*"
python-dotenv = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
